featt: add templating and exceptions
